Move tests for timing-related triggers to their own module
