+++ tests/test_cases/test_cocotb/test_cocotb.py
@@ -51,18 +51,6 @@ from cocotb.utils import get_sim_time
 from cocotb.outcomes import Value, Error
 
 
-@cocotb.test(expect_fail=False)
-def test_function_reentrant_clock(dut):
-    """Test yielding a reentrant clock"""
-    clock = dut.clk
-    timer = Timer(100)
-    for i in range(10):
-        clock <= 0
-        yield timer
-        clock <= 1
-        yield timer
-
-
 @cocotb.coroutine
 def clock_gen(clock):
     """Example clock gen for test use"""
@@ -157,59 +145,6 @@ def test_clock_with_units(dut):
 
     clk_gen.kill()
 
-@cocotb.test(expect_fail=False)
-def test_timer_with_units(dut):
-    time_fs = get_sim_time(units='fs')
-
-    # Yield for one simulation time step
-    yield Timer(1)
-    time_step = get_sim_time(units='fs') - time_fs
-
-    try:
-        # Yield for 2.5 timesteps, should throw exception
-        yield Timer(2.5*time_step, units='fs')
-        raise TestFailure("Timers should throw exception if time cannot be achieved with simulator resolution")
-    except ValueError:
-        dut._log.info("As expected, unable to create a timer of 2.5 simulator time steps")
-
-    time_fs = get_sim_time(units='fs')
-
-    yield Timer(3, "ns")
-
-    if get_sim_time(units='fs') != time_fs+3000000.0:
-        raise TestFailure("Expected a delay of 3 ns")
-
-    time_fs = get_sim_time(units='fs')
-    yield Timer(1.5, "ns")
-
-    if get_sim_time(units='fs') != time_fs+1500000.0:
-        raise TestFailure("Expected a delay of 1.5 ns")
-
-    time_fs = get_sim_time(units='fs')
-    yield Timer(10.0, "ps")
-
-    if get_sim_time(units='fs') != time_fs+10000.0:
-        raise TestFailure("Expected a delay of 10 ps")
-
-    time_fs = get_sim_time(units='fs')
-    yield Timer(1.0, "us")
-
-    if get_sim_time(units='fs') != time_fs+1000000000.0:
-        raise TestFailure("Expected a delay of 1 us")
-
-@cocotb.test()
-def test_timer_with_rational_units(dut):
-    """ Test that rounding errors are not introduced in exact values """
-    # now with fractions
-    time_fs = get_sim_time(units='fs')
-    yield Timer(Fraction(1, int(1e9)), units='sec')
-    assert get_sim_time(units='fs') == time_fs + 1000000.0, "Expected a delay of 1 ns"
-
-    # now with decimals
-    time_fs = get_sim_time(units='fs')
-    yield Timer(Decimal('1e-9'), units='sec')
-    assert get_sim_time(units='fs') == time_fs + 1000000.0, "Expected a delay of 1 ns"
-
 
 @cocotb.test(expect_fail=False)
 def test_anternal_clock(dut):
@@ -221,98 +156,6 @@ def test_anternal_clock(dut):
         yield RisingEdge(dut.clk)
         count += 1
     clk_gen.kill()
-
-exited = False
-
-
-@cocotb.coroutine
-def do_test_readwrite_in_readonly(dut):
-    global exited
-    yield RisingEdge(dut.clk)
-    yield ReadOnly()
-    yield ReadWrite()
-    exited = True
-
-
-@cocotb.coroutine
-def do_test_cached_write_in_readonly(dut):
-    global exited
-    yield RisingEdge(dut.clk)
-    yield ReadOnly()
-    dut.clk <= 0
-    exited = True
-
-
-@cocotb.coroutine
-def do_test_afterdelay_in_readonly(dut, delay):
-    global exited
-    yield RisingEdge(dut.clk)
-    yield ReadOnly()
-    yield Timer(delay)
-    exited = True
-
-
-@cocotb.test(expect_error=True,
-             expect_fail=cocotb.SIM_NAME.lower().startswith(("icarus",
-                                                             "riviera",
-                                                             "modelsim",
-                                                             "ncsim",
-                                                             "xmsim")))
-def test_readwrite_in_readonly(dut):
-    """Test doing invalid sim operation"""
-    global exited
-    exited = False
-    clk_gen = cocotb.fork(Clock(dut.clk, 100).start())
-    coro = cocotb.fork(do_test_readwrite_in_readonly(dut))
-    yield [Join(coro), Timer(10000)]
-    clk_gen.kill()
-    if exited is not True:
-        raise TestFailure
-
-@cocotb.test(expect_error=True,
-             expect_fail=cocotb.SIM_NAME.lower().startswith(("icarus",
-                                                             "riviera",
-                                                             "modelsim",
-                                                             "ncsim",
-                                                             "xmsim")))
-def test_cached_write_in_readonly(dut):
-    """Test doing invalid sim operation"""
-    global exited
-    exited = False
-    clk_gen = cocotb.fork(Clock(dut.clk, 100).start())
-    coro = cocotb.fork(do_test_cached_write_in_readonly(dut))
-    yield [Join(coro), Timer(10000)]
-    clk_gen.kill()
-    if exited is not True:
-        raise TestFailure
-
-
-@cocotb.test(expect_fail=cocotb.SIM_NAME.lower().startswith(("icarus",
-                                                             "chronologic simulation vcs")),
-             skip=cocotb.SIM_NAME.lower().startswith(("ncsim", "xmsim")))
-def test_afterdelay_in_readonly(dut):
-    """Test doing invalid sim operation"""
-    global exited
-    exited = False
-    clk_gen = cocotb.fork(Clock(dut.clk, 100).start())
-    coro = cocotb.fork(do_test_afterdelay_in_readonly(dut, 0))
-    yield [Join(coro), Timer(1000)]
-    clk_gen.kill()
-    if exited is not True:
-        raise TestFailure
-
-
-@cocotb.test()
-def test_afterdelay_in_readonly_valid(dut):
-    """Same as test_afterdelay_in_readonly but with valid delay > 0"""
-    global exited
-    exited = False
-    clk_gen = cocotb.fork(Clock(dut.clk, 100).start())
-    coro = cocotb.fork(do_test_afterdelay_in_readonly(dut, 1))
-    yield [Join(coro), Timer(100000)]
-    clk_gen.kill()
-    if exited is not True:
-        raise TestFailure
 
 
 @cocotb.coroutine
@@ -734,38 +577,6 @@ def test_clock_cycles_forked(dut):
 
 
 @cocotb.test()
-def test_readwrite(dut):
-    """ Test that ReadWrite can be waited on """
-    # gh-759
-    yield Timer(1)
-    dut.clk <= 1
-    yield ReadWrite()
-
-
-@cocotb.test()
-def test_writes_have_taken_effect_after_readwrite(dut):
-    """ Test that ReadWrite fires first for the background write coro """
-    dut.stream_in_data.setimmediatevalue(0)
-
-    @cocotb.coroutine
-    def write_manually():
-        yield ReadWrite()
-        # this should overwrite the write written below
-        dut.stream_in_data.setimmediatevalue(2)
-
-    # queue a backround task to do a manual write
-    waiter = cocotb.fork(write_manually())
-
-    # do a delayed write. This will be overwritten
-    dut.stream_in_data <= 3
-    yield waiter
-
-    # check that the write we expected took precedence
-    yield ReadOnly()
-    assert dut.stream_in_data.value == 2
-
-
-@cocotb.test()
 def test_trigger_with_failing_prime(dut):
     """ Test that a trigger failing to prime throws """
     class ABadTrigger(Trigger):
@@ -779,28 +590,6 @@ def test_trigger_with_failing_prime(dut):
         assert "oops" in str(exc)
     else:
         raise TestFailure
-
-
-@cocotb.coroutine
-def example():
-    yield Timer(10, 'ns')
-    return 1
-
-
-@cocotb.test()
-def test_timeout_func_fail(dut):
-    try:
-        yield cocotb.triggers.with_timeout(example(), timeout_time=1, timeout_unit='ns')
-    except cocotb.result.SimTimeoutError:
-        pass
-    else:
-        assert False, "Expected a Timeout"
-
-
-@cocotb.test()
-def test_timeout_func_pass(dut):
-    res = yield cocotb.triggers.with_timeout(example(), timeout_time=100, timeout_unit='ns')
-    assert res == 1
 
 
 @cocotb.test()
